feat: overload `hash` returns based on `format`
The original typings just simply signify that the return is either a
Uint8Array or a string regardless of the desired format.

This meant that all callers of `hash` must perform a cast every time
`hash` is called. This change will have the type system guarantee the
already existing behavior.